App debugueada y ejecutable, necesario tener gradle 8.0 y todas dependencia actualizadas, especialmente tools android de google... problema de ui en app

+++ app/src/main/java/com/example/movieapp/data/model/Movie.kt
@@ -13,7 +13,7 @@ data class Movie(
     val relese_date: String = "",
     val title: String = "",
     val video: Boolean = false,
-    val vote_average: Int = -1,
+    val vote_average: Double = -1.0,
     val vote_count: Int = -1,
 )
